+++ argopy/stores/fsspec_wrappers.py
@@ -4,9 +4,10 @@ import xarray as xr
 import pandas as pd
 import requests
 import fsspec
-import pickle
-import tempfile
 import shutil
+import pickle
+import json
+import tempfile
 from IPython.core.display import display, HTML
 
 from argopy.options import OPTIONS
@@ -14,18 +15,11 @@ from argopy.errors import ErddapServerError, FileSystemHasNoCache, CacheFileNotF
 from abc import ABC, abstractmethod
 
 
-class argo_store_proto(ABC):
-    """ Provide fsspec like file systems
-
-        Simply use a ``cache`` option to implement a filecache file system
-        Only a few methods from fsspec are provided like "open" or "glob"
-        All open URI are registered to implement a clear cache method not available from fsspec
-    """
-
+class argo_store_proto(ABC):  # Should this class inherits from fsspec.spec.AbstractFileSystem ?
     protocol = ''  # One in fsspec.registry.known_implementations
 
     def __init__(self, cache: bool = False, cachedir: str = "", **kw):
-        """ Create a file storage system
+        """ Create a file storage system for Argo data
 
             Parameters
             ----------
@@ -47,12 +41,15 @@ class argo_store_proto(ABC):
             # since this is the update frequency of the Ifremer erddap
             self.cache_registry = []  # Will hold uri cached by this store instance
 
-    def open(self, url, *args, **kwargs):
-        self.register(url)
-        return self.fs.open(url, *args, **kwargs)
+    def open(self, path, *args, **kwargs):
+        self.register(path)
+        return self.fs.open(path, *args, **kwargs)
 
     def glob(self, path, **kwargs):
         return self.fs.glob(path, **kwargs)
+
+    def exists(self, path, *args):
+        return self.fs.exists(path, *args)
 
     def store_path(self, uri):
         if not uri.startswith(self.fs.target_protocol):
@@ -117,7 +114,7 @@ class argo_store_proto(ABC):
 
 
 class filestore(argo_store_proto):
-    """ Wrapper around fsspec file stores
+    """Wrapper around fsspec file stores
 
         https://filesystem-spec.readthedocs.io/en/latest/api.html#fsspec.implementations.local.LocalFileSystem
 
@@ -160,12 +157,10 @@ class filestore(argo_store_proto):
 
 
 class httpstore(argo_store_proto):
-    """ Wrapper around fsspec http file store
+    """Wrapper around fsspec http file store
 
-        This wrapper:
-         - intend to make argopy safer to failures from http requests
-         - is primarily used by the Erddap data/index fetchers
-         - provide more verbose errors from http requests
+        This wrapper intend to make argopy safer to failures from http requests
+        This wrapper is primarily used by the Erddap data/index fetchers
     """
     protocol = "http"
 
@@ -218,6 +213,26 @@ class httpstore(argo_store_proto):
             error.append("The URL triggering this error was: \n%s" % url)
             print("\n".join(error))
             r.raise_for_status()
+
+    def open_json(self, url, **kwargs):
+        """ Return a json from an url, or verbose errors
+
+            Parameters
+            ----------
+            url: str
+
+            Returns
+            -------
+            json
+
+        """
+        try:
+            with self.fs.open(url) as of:
+                js = json.load(of, **kwargs)
+            self.register(url)
+            return js
+        except requests.HTTPError as e:
+            self._verbose_exceptions(e)
 
     def open_dataset(self, url, **kwargs):
         """ Return a xarray.dataset from an url, or verbose errors