+++ argopy/tests/test_utilities.py
@@ -3,8 +3,12 @@ import io
 import pytest
 import unittest
 import argopy
+import xarray as xr
+import numpy as np
+import requests
+import shutil
 from argopy.utilities import load_dict, mapp_dict, list_multiprofile_file_variables, \
-    isconnected, erddap_ds_exists, list_available_data_src
+    isconnected, erddap_ds_exists, open_etopo1, list_available_data_src, linear_interpolation_remap
 from argopy import DataFetcher as ArgoDataFetcher
 
 AVAILABLE_SOURCES = list_available_data_src()
@@ -64,3 +68,37 @@ def test_clear_cache():
 #         assert isinstance(ds, xr.DataArray) is True
 #     except requests.HTTPError:  # not our fault
 #         pass
+
+
+class test_linear_interpolation_remap(unittest.TestCase):
+    @pytest.fixture(autouse=True)
+    def create_data(self):
+        # create fake data to test interpolation:
+        temp = np.random.rand(200, 100)
+        pres = np.sort(np.floor(np.zeros(
+            [200, 100])+np.linspace(50, 950, 100)+np.random.randint(-5, 5, [200, 100])))
+        self.dsfake = xr.Dataset({'TEMP': (['N_PROF', 'N_LEVELS'], temp),
+                                  'PRES': (['N_PROF', 'N_LEVELS'],  pres)},
+                                 coords={'N_PROF': ('N_PROF', range(200)),
+                                         'N_LEVELS': ('N_LEVELS', range(100)),
+                                         'Z_LEVELS': ('Z_LEVELS', np.arange(100, 900, 20))})
+
+    def test_interpolation(self):
+        # Run it with success:
+        dsi = linear_interpolation_remap(
+            self.dsfake.PRES, self.dsfake['TEMP'], self.dsfake['Z_LEVELS'], z_dim='N_LEVELS', z_regridded_dim='Z_LEVELS')
+        assert 'remapped' in dsi.dims
+
+    def test_error_zdim(self):
+        # Test error:
+        # catches error from _regular_interp linked to z_dim
+        with pytest.raises(RuntimeError):
+            dsi = linear_interpolation_remap(
+                self.dsfake.PRES, self.dsfake['TEMP'], self.dsfake['Z_LEVELS'], z_regridded_dim='Z_LEVELS')
+
+    def test_error_ds(self):
+        # Test error:
+        # catches error from linear_interpolation_remap linked to datatype
+        with pytest.raises(ValueError):
+            dsi = linear_interpolation_remap(
+                self.dsfake.PRES, self.dsfake, self.dsfake['Z_LEVELS'], z_dim='N_LEVELS', z_regridded_dim='Z_LEVELS')