+++ docs/conf.py
@@ -21,12 +21,12 @@ import datetime
 from contextlib import suppress
 allowed_failures = set()
 
-if 'conda' in sys.executable:
-    print('conda environment:')
-    subprocess.run(['conda', 'list'])
-else:
-    print('pip environment:')
-    subprocess.run(['pip', 'list'])
+# if 'conda' in sys.executable:
+#     print('conda environment:')
+#     subprocess.run(['conda', 'list'])
+# else:
+#     print('pip environment:')
+#     subprocess.run(['pip', 'list'])
 
 print("xarray: %s, %s" % (xarray.__version__, xarray.__file__))